Extract lens library handlers
