+++ tests/test-correctness-phase2.js
@@ -193,6 +193,7 @@ const pwaAppShellAssets = [
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
   '/js/lens.js',
+  '/js/lens-library.js',
   '/js/lens-local.js',
   '/js/lens-local-worker.js',
   '/js/lens-local-utils.js',
